# Src/meetDoc/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MeetDoc",
        "description": "MeetDoc APIs Documentation using Swagger",
        "contact": {
            "email": "mohamed.ahmed.aly.d@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notes/{id}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get Patient Notes by patient id",
                "operationId": "3620e5d93c58d029e1ba771224ad00d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Patient",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Patient Symptoms returned successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/add-note": {
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Add note to patient",
                "operationId": "91446779c8945450a577e0806e1a4972",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "patient_id": {
                                        "type": "number"
                                    },
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "note has been added",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/symptoms/{id}": {
            "get": {
                "tags": [
                    "Symptoms"
                ],
                "summary": "Get Patient Symptoms by patient id",
                "operationId": "0e45fbde3a6fb193f70684ef99886aca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Patient",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Patient Symptoms returned successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/logout": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Logout Authnticated user by removing it's token",
                "operationId": "Logout",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Login",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "device_id": {
                                        "type": "string"
                                    },
                                    "new_device": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}